Refactored functions tests.
